Cleaned up and improved basic character improvement. Also added a working 3rd person camera
Basically simplified the old player movement code and added smooth rotation so the player looks in the direction they are moving. Decided to get rid of the old 3rd person camera I was working on and added a new one using the Unity package Cinemachine. This mainly saves me a lot of time and move on to other things for development